Switch between light and dark theme depending on time of day
It may seem backwards, but I used a light terminal at night because web
pages tend to be bright and switching to a bright web page after using
the terminal is harsh.

To combat this I use Night Shift on the mac which means I can handle the
bright colours.

TODO: make this take into account local sunrise and sunset.